+++ app/src/main/kotlin/io/sakurasou/model/entity/Group.kt
@@ -12,5 +12,5 @@ data class Group(
     val name: String,
     val description: String?,
     val strategyId: Long,
-    val maxSize: Double = 5 * 1024 * 1024 * 1024.0
+    val maxSize: Long = 5 * 1024 * 1024 * 1024L
 )
